+++ tests/unit/base/test_inplace.py
@@ -20,14 +20,17 @@ Note:
 
 import os
 import pathlib
+from importlib.util import find_spec
 from unittest.mock import MagicMock, patch
 
 import pytest
 
+from model_navigator.exceptions import ModelNavigatorUserInputError
 from model_navigator.inplace.config import DEFAULT_CACHE_DIR, OptimizeConfig, inplace_cache_dir
 from model_navigator.inplace.model import OptimizedModule, PassthroughModule, RecordModule
 from model_navigator.inplace.registry import module_registry
 from model_navigator.inplace.utils import get_object_name
+from model_navigator.inplace.wrapper import Module
 
 
 def test_get_object_name():
@@ -137,3 +140,60 @@ def test_recording_model_is_ready_for_optimization_returns_true_when_enough_samp
     )
     module._samples_shapes = {"hash": list(range(inplace_config.min_num_samples))}
     assert module.is_ready_for_optimization
+
+
+@pytest.mark.skipif(not find_spec("torch"), reason="PyTorch is not installed.")
+def test_record_module_alt_forward():
+    import torch  # pytype: disable=import-error
+
+    module = MagicMock()
+    module.side_effect = lambda _: torch.Tensor([1])
+    module.forward.side_effect = lambda _: torch.Tensor([2])
+
+    module1 = RecordModule(module=module, name="model1", input_mapping=lambda x: x, output_mapping=lambda x: x)
+    assert module1(torch.Tensor([0])) == torch.ones(1)
+
+    module1 = RecordModule(
+        module=module, name="model2", input_mapping=lambda x: x, output_mapping=lambda x: x, forward=module.forward
+    )
+    assert module1(torch.Tensor([0])) == torch.Tensor([2])
+
+
+@pytest.mark.skipif(not find_spec("torch"), reason="PyTorch is not installed.")
+def test_module_wrapper_alt_forward():
+    import torch  # pytype: disable=import-error
+
+    module = MagicMock()
+    # torch __call__ triggers forward, so we need to mock both, and be careful
+    module.side_effect = lambda _: module.forward(0) + 1000
+    module.forward.side_effect = lambda _: torch.Tensor([2])
+
+    module = Module(module, "model3", forward_func="forward")
+
+    assert module(torch.zeros(1)) == torch.Tensor([2])
+
+
+@pytest.mark.skipif(not find_spec("torch"), reason="PyTorch is not installed.")
+def test_module_wrapper_torch_object():
+    import torch  # pytype: disable=import-error
+
+    class TestModule(torch.nn.Module):
+        def forward(self, x):
+            return x + 1
+
+    module = TestModule()
+    nav_module = Module(module, "model4", forward_func="forward")
+
+    assert nav_module(torch.Tensor([1000])) == torch.Tensor([1001])
+
+
+@pytest.mark.skipif(not find_spec("torch"), reason="PyTorch is not installed.")
+def test_module_wrapper_complain_on_missing_custom_func_name():
+    import torch  # pytype: disable=import-error
+
+    class TestModule(torch.nn.Module):
+        def forward(self, x):
+            return x + 1
+
+    with pytest.raises(ModelNavigatorUserInputError):
+        Module(TestModule(), "model5", forward_func="non_existing_func")